Add exercise tracking data to Exercise class and move to correct package
